+++ src/main/java/com/supermartijn642/pottery/PotteryClient.java
@@ -7,7 +7,6 @@ import com.supermartijn642.pottery.content.PotBlockRenderer;
 import com.supermartijn642.pottery.content.PotColor;
 import com.supermartijn642.pottery.content.PotType;
 import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.world.level.block.entity.DecoratedPotPatterns;
 
 /**
  * Created 27/11/2023 by SuperMartijn642
@@ -28,8 +27,9 @@ public class PotteryClient {
             }
         }
         // Put all decorated pot patterns into the block atlas
-        BuiltInRegistries.DECORATED_POT_PATTERNS.registryKeySet().stream()
-            .map(DecoratedPotPatterns::location)
+        BuiltInRegistries.DECORATED_POT_PATTERN.holders()
+            .filter(holder -> holder.key().location().getNamespace().equals("minecraft"))
+            .map(holder -> holder.value().assetId().withPrefix("entity/decorated_pot/"))
             .forEach(texture -> handler.registerAtlasSprite(TextureAtlases.getBlocks(), texture));
     }
 }