+++ src/main/java/com/supermartijn642/pottery/content/PotColor.java
@@ -1,11 +1,13 @@
 package com.supermartijn642.pottery.content;
 
 import com.supermartijn642.pottery.Pottery;
+import net.minecraft.client.renderer.Sheets;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.world.level.block.entity.DecoratedPotPattern;
 import net.minecraft.world.level.block.entity.DecoratedPotPatterns;
 
 import java.util.Locale;
@@ -71,13 +73,13 @@ public enum PotColor {
         return this.dyeIngredient.get();
     }
 
-    public ResourceLocation getPatternLocation(ResourceKey<String> key){
-        ResourceLocation texture = DecoratedPotPatterns.location(key);
+    public ResourceLocation getPatternLocation(ResourceKey<DecoratedPotPattern> key){
+        ResourceLocation texture = Sheets.DECORATED_POT_MATERIALS.get(key).texture();
         if(this == BLANK)
             return texture;
 
         if(key.location().getNamespace().equals("minecraft"))
-            return new ResourceLocation(Pottery.MODID, "patterns/" + this.getIdentifier() + "/" + texture.getPath().substring(texture.getPath().lastIndexOf('/') + 1));
+            return ResourceLocation.fromNamespaceAndPath(Pottery.MODID, "patterns/" + this.getIdentifier() + "/" + texture.getPath().substring(texture.getPath().lastIndexOf('/') + 1));
 
         return texture;
     }