+++ changelog.md
@@ -1,2 +1,2 @@
 ### Pottery 1.0.1
-- Initial release of Pottery for Minecraft 1.20.5 & 1.20.6
+- Initial release of Pottery for Minecraft 1.21 & 1.21.1